Split RL actions package into `base` and `registry` modules
