+++ src/mqt/predictor/rl/actions/__init__.py
@@ -8,138 +8,17 @@
 
 """Actions (i.e. compiler passes) available in the reinforcement learning environment."""
 
-# ruff: file-ignore[non-empty-init-module]
-
 from __future__ import annotations
 
-from collections import defaultdict
-from dataclasses import dataclass
-from enum import Enum
-from typing import TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    from collections.abc import Callable
-
-    from qiskit.passmanager import PropertySet
-
-
-class CompilationOrigin(str, Enum):
-    """Enumeration of the origin of the compilation action."""
-
-    QISKIT = "qiskit"
-    TKET = "tket"
-    BQSKIT = "bqskit"
-
-
-class PassType(str, Enum):
-    """Enumeration of the type of compilation pass."""
-
-    OPT = "optimization"
-    SYNTHESIS = "synthesis"
-    MAPPING = "mapping"
-    LAYOUT = "layout"
-    ROUTING = "routing"
-    FINAL_OPT = "final_optimization"
-    TERMINATE = "terminate"
-
-
-@dataclass
-class Action:
-    """Base class for all actions in the reinforcement learning environment.
-
-    Attributes:
-        name: Unique action name.
-        origin: SDK origin of the action; ``None`` for terminate.
-        pass_type: Category of pass represented by this action.
-        transpile_pass: Pass object(s) executed for this action.
-        preserves_layout: Whether action preserves existing layout.
-        preserves_routing: Whether action preserves existing routing.
-        preserves_synthesis: Whether action preserves synthesis state.
-    """
-
-    name: str
-    origin: CompilationOrigin | None
-    pass_type: PassType
-    transpile_pass: Any
-    preserves_layout: bool = False
-    preserves_routing: bool = False
-    preserves_synthesis: bool = False
-
-
-@dataclass
-class DeviceIndependentAction(Action):
-    """Action that represents a static compilation pass that can be applied directly."""
-
-
-@dataclass
-class DeferredDeviceAction(Action):
-    """Action that defers construction of a device-specific pass.
-
-    Attributes:
-        do_while: Optional do-while predicate for pass-manager execution.
-    """
-
-    transpile_pass: Any
-    do_while: Callable[[PropertySet], bool] | None = None
-
-
-_ACTIONS: dict[str, Action] = {}
-
-
-def register_action(action: Action) -> Action:
-    """Registers a new Action in the global _ACTIONS registry.
-
-    Args:
-        action: Action to register.
-
-    Returns:
-        The registered Action.
-
-    Raises:
-        ValueError: If an action with the same name is already registered.
-    """
-    if action.name in _ACTIONS:
-        msg = f"Action with name {action.name} already registered."
-        raise ValueError(msg)
-    _ACTIONS[action.name] = action
-    return action
-
-
-def get_actions_by_pass_type() -> dict[PassType, list[Action]]:
-    """Groups registered Actions from the global _ACTIONS registry by PassType.
-
-    Returns:
-        A dictionary mapping each PassType to the list of registered Actions of that type.
-    """
-    result: dict[PassType, list[Action]] = defaultdict(list)
-    for action in _ACTIONS.values():
-        result[action.pass_type].append(action)
-    return result
-
-
 from mqt.predictor.rl.actions import bqskit_actions, qiskit_actions, tket_actions
-
-for _action in (
-    *qiskit_actions.qiskit_layout_actions(),
-    qiskit_actions.qiskit_mapping_action(),
-    qiskit_actions.qiskit_synthesis_action(),
-    qiskit_actions.qiskit_o3_action(),
-    *qiskit_actions.qiskit_optimization_actions(),
-    qiskit_actions.qiskit_final_optimization_action(),
-    tket_actions.tket_routing_action(),
-    *tket_actions.tket_optimization_actions(),
-    *bqskit_actions.bqskit_layout_actions(),
-    *bqskit_actions.bqskit_routing_actions(),
-    *bqskit_actions.bqskit_mapping_actions(),
-    *bqskit_actions.bqskit_synthesis_actions(),
-    DeviceIndependentAction(
-        "terminate",
-        None,
-        PassType.TERMINATE,
-        transpile_pass=[],
-    ),
-):
-    register_action(_action)
+from mqt.predictor.rl.actions.base import (
+    Action,
+    CompilationOrigin,
+    DeferredDeviceAction,
+    DeviceIndependentAction,
+    PassType,
+)
+from mqt.predictor.rl.actions.registry import get_actions_by_pass_type, register_action
 
 __all__ = [
     "Action",
